requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
asyncpg

sqlalchemy
psycopg[binary]
alembic
langgraph-checkpoint-postgres

mistralai
httpx

langchain
langchain-core
langchain_mistralai
langchain-community
langgraph
mcp[cli]
fastmcp
mcp

requests
tenacity
beautiful-soup 
tavily-python
google-search-results
firecrawl-py

pgvector
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langgraph-checkpoint-postgres
pytest
celery 
redis